refactor soccer game application
